Teach the tracker to read the Python stack when tracking allocations
Reading the Python stack from the allocation functions is too expensive
as the GIL needs to be held when using the C-API and this not only
serializes the whole program on allocations but also can deadlock
extension modules that keep locks that are not GIL-aware.

For this reason, we need to maintain a per-thread data structure that
allows unsynchronized access to the Python stack frame. To achieve this,
we install a native profile function that maintains a mirror data
structure of the Python stack frame that has the required properties and
that can be accessed from the allocation tracking functions.

To keep the access to the global state more organized, we also add a new
class that manages all global state and offers the interface to the
tracking machinery.

Authored-by: Pablo Galindo <pgalindo3@bloomberg.net>
Reviewed-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Signed-off-by: Pablo Galindo <pgalindo3@bloomberg.net>