Add a `Helgrind` step and fix unsynchronized code
To protect ourselves against data races and the like, we can add a new
step that runs `Helgrind` in the CI to get notified of any possible
data race. This requires to fix some false positives and add some extra
suppression for code that `Helgrind` cannot see (like the atomic
boolean).