Save results into a file
The serialization and de-serialization is implemented with a simple text
format where each line represents either a frame action sequence
(containing a key to the frame itself and whether it's a push or pop),
an allocation event or a frame.

We keep track of the frames in a map and hash their contents to ensure
we can uniquely identify and de-duplicate them. We write this map of
frames at the end of the tracking process. The other data structures are
streamed to the output buffer as we see them to avoid keeping large
buffers in memory.

Note that this serializer writes directly into the output stream instead
of relying on the `RecordWriter` and background threads in order to
reduce the size of the patch. The performance is already quite good with
serial writes but we can further improve this by refactoring the
`RecordWriter` to handle the new serialization types.