Keep track of stacks on each thread
The previous implementation of the frame tracking code did not account
for threads, so we were never able to resolve the stack traces for the
allocations which happened on threads other than the main thread.

This change logs the thread ID as part of the stack sequence and uses
these to reconstruct multiple stack traces - one per thread - in the
parser.