Use a binary writer to speed up the I/O
Turns out streams operators are quite slow and we are using the heavily
in the hottest part of the tracking (the allocation/deallocation)
tracking. To improve our throughput this commit does the following:

* Substitute stream operators to bare write() calls.
* Use a binary format instead of text.
* Implement our own buffering system to minimize the system calls.

Using a binary format can make debugging a bit more challenging but we
can implement an alternative text-based format for debugging purposes
in the future.