Speed up traceback generation by caching Python strings
Is quite common that the strings representing the function names and the
file names of the stack frames are repeated, as many allocations tend to
happen on the same location in the stack. Producing independent Python
strings for every repeated case of these strings is not only wasteful
but also time intensive. To improve the memory consumption of the
program and the time spent generating the tracebacks, we can provide a
cache of Python strings to avoid generating the same string twice. This
is still better than interning them because we also avoid *generating*
the string object in the first place if it has been previously produced.