Correctly populate line numbers in the tracebacks
We are using `sys.setprofile()` to making the tracking efficient (so
individual lines are not traced), and currently, we are gathering the
filename, the function name and the line number. But as
`sys.setprofile()` only records function starts and returns, the line
number will always correspond to the entry point and the return point,
which we already know because we know where the function is defined.
There is no value to recording these numbers, at least as we are doing
things right now.

To solve this we store the parent line number (so we know from where the
function is being called) on every frame PUSH and POP. This solves the
problem for all the frames except the one where the allocation happens.
For the one in which the allocation happens, a new thread local pointer
variable is added so that we can read from the tracking allocation
functions in a way we can retrieve the current line number at the time
the allocation happens. This allows us to skip having a trace function,
which is a no go because we cannot write line-by-line information to the
file.