Generate truly unique identifiers per frame item
When parsing records we are not generating truly unique identifiers for
every frame due to the fact that all frames associated with allocations
that happen in the same function have the same frame identifier as the
line number is not ready until the allocation happens and that is the
only thing that can distinguish them.

Having unique identifiers per frame is specially useful since in this
way we can identify a whole stack trace by a single integer, which not
only helps in performance but also simplifies things considerably in the
reporter.