Add API to get a high-watermark snapshot
Add an API that allows to retrieve the records that form the heap state
at the moment the program reached the high memory watermark (the moment
the program was using the most memory). The collection of records will
be aggregated by the stack trace, so all allocations that happen in the
exact same place will be reported as a single allocation with the
cumulative size (this makes iterating over all allocations faster and
reduces the memory footprint and the post-processing work in Python).