Use a single function to track allocations and deallocations
Some deallocators (such as `munmap`) can do partial deallocations so is
not true that the size of the deallocations must be always 0. By using a
single API for tracking both allocations and deallocations we can
simplify the code a bit, as deallocations still need to be identified by
looking at the "allocator" field (at least for now).

Authored-by: Pablo Galindo <pgalindo3@bloomberg.net>
Reviewed-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Reviewed-by: Pradyun Gedam <pgedam@bloomberg.net>
Signed-off-by: Pablo Galindo <pgalindo3@bloomberg.net>