Improve the tracking for `realloc()`
Using the same "allocator" type for both the virtual allocation and the
deallocation in realloc() is not super useful because then REALLOC as a
value for the enum can be both an allocator or a deallocator. This
commit changes it so we get a FREE for the virtual deallocation and
REALLOC for the allocation.

Authored-by: Pablo Galindo <pgalindo3@bloomberg.net>
Reviewed-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Signed-off-by: Pablo Galindo <pgalindo3@bloomberg.net>