Refactor the `StackTraceTree` class
Prepare the `StackTraceTree` to be used from the tracking machinery as
well by moving it into its own header file and by allowing it to receive
a callback to be executed when a new frame is inserted in the tree. This
will make the class more versatile and will allow us to invoke the
record writer in a similar way as the one we currently use for handling
Python frames.

Authored-by: Pablo Galindo <pgalindo3@bloomberg.net>
Reviewed-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Signed-off-by: Pablo Galindo <pgalindo3@bloomberg.net>