Add `pytest` markers to select test to run in `Valgrind`
To avoid having to keep a separate list in the Makefile of the tests
that should run under `Valgrind` and to allow more easily to select
individual tests instead of entire files, use `pytest` markers to mark
the tests that we want to run under `Valgrind`. Then, to run only that
subset of tests one can do `python -m pytest -m valgrind`.

Authored-by: Pablo Galindo <pgalindo3@bloomberg.net>
Reviewed-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Signed-off-by: Pablo Galindo <pgalindo3@bloomberg.net>