Breakout sub-commands into a commands subpackage
This also establishes an interface for adding additional sub-commands,
which allows the code for each subcommand to be clearly separated into
different sub-modules.

Also, add a slash to the epilog to make all lines dedent correctly.

Also, change how main handles arguments which moves slicing of
`sys.argv` to be independent of CLI calls.

Authored-by: Pradyun Gedam <pgedam@bloomberg.net>
Reviewed-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Reviewed-by: Pablo Galindo <pgalindo3@bloomberg.net>
Signed-off-by: Pradyun Gedam <pgedam@bloomberg.net>