Refactor the allocation parsing functions in the record reader
There is some duplication in the functions that are parsing all
allocation records in the record reader. Currently one function returns
element by element (transformed to Python objects) so we can use it as a
generator while the other returns a vector of those elements so we can
feed that into the function that calculates the high watermark snapshot.

By refactoring this functions to use a common central piece that parses
record by record but do not transform them immediately into Python
objects, we can eliminate this duplication and move the collection logic
to `Cython` so we can use progress bars more easily in the future.