Fix `filter_relevant_allocations` to correctly handle records
The `filter_relevant_allocations` function can still pick up incorrect
records as there is the possibility that there are `free` calls that are
deallocating something with an address that is then reused for some of
the allocations we care about. In this case, the call to
`filter_relevant_allocations` will give records for `free` that don't
correspond to the `valloc` (or whatever we want) that we search for. The
correct implementation is to iterate over the records keeping track of
the allocations and then only report the frees if we have seen an
allocation with that address, otherwise, they should be ignored.