Add the `symbolizer` machinery
Add a new collection of utilities that allows to efficiently obtain the
symbol name, file name and line number associated to a given instruction
pointer. For the `symbolizer` to work, it needs to be feeded with the
different states of the memory segments associated to the program. The
`symbolizer` needs to be also informed every time this state changes.