Teach the parser to read native frame information
This commit adds the native tracking functionality to the parser so it
can successfully reconstruct the information associated to the native
frames. This involves fetching the `symbolizer` with the memory segment
information every time the parser registers that a change in the memory
maps has been written to the file in addition to reconstructing the
graph of native frames so they can be reconstructed from the initial
native index that is now stored in the allocation records.