Fail for executable paths longer than PATH_MAX
PATH_MAX is a misnomer - this is in no way an actual enforced maximum,
and it's entirely possible for the path to an executable to be longer
than this. The previous code would fail to detect this case and would
truncate the executable path.

Detect this case and raise an exception if it occurs.

Also - I'm changing `execname` to `executable` since this isn't a name,
it's a path.