Reverse the iteration direction of the NativeTrace
For the Python frames, we are returning an iterable that starts with the
symbol closer to where the allocation happen but `libunwind` fills
the buffer where it places the instruction pointers with the symbol
closer to the allocation being at the start. This makes the output the
reverse of what we are returning for the Python-based stack trace.

To fix this, the easiest way is to implement a simple reverse iterator
for the `NativeTrace` class.