Show the number of allocations on the flame graph
It can appear confusing that every frame is only shown once on the flame
graph. By displaying the number of allocations occurring in each frame,
we can make it more clear there could have been several calls to the
allocating frame.

Note that parent frames sum the childrens' number of allocations, just
like with the amount of total allocated memory.