Some `Cython` frames have incorrect line numbers
`Cython` frames have some incorrect line numbers in some occasions.
This is happening because `Cython` creates new frames so the tracing
functions can work as if those were regular Python frames. The frame
constructor then inspects the value of the interpreter state to populate
the value of `frame.f_back` (which is supposed to point to a previous
frame). This value will *not* be a `Cython` frame because those are not
registered in the interpreter, so all `Cython` frames that we receive in
the tracking function have an incorrect value for `frame.f_back` that
points to the last pure Python frame.

This manifest in incorrect line numbers because we correctly reconstruct
the stack as the mechanism that we employ is based on PUSH/POP
operations, but we rely on a pointer to the current frame to populate
the current line number when the allocation happens. As we use the value
of `frame.f_back` to update the current frame pointer, then the line
numbers are wrong.

The easier way to solve this is to keep a copy of the stack manually and
update it as we do the PUSH/POP operations and use that instead.