Rename "tooltip" field to "location" in flamegraph
We now generate two fragments for the tooltip display on the Python
side: the location information (which comes first) and the number of
allocations, which is at the end of the tooltip. The middle part is
generated in the JS code. Naming the `tooltip` field `location` makes it
clear that we are not generating the whole tooltip in Python and the
frontend is assembling it from various labels.