Fix crash if allocator functions are called with no Tracker instance
As checking if the tracker is active or not requires having a tracker
instance alive, there are many scenarios where we will crash because the
tracker instance is gone but the tracking functions are still referenced
somehow. These scenarios can happen if:

* A thread that was created during tracking calls the trace function
  once tracking has finished.
* There are some references to the allocation function held in some
  function pointer.

We need to ensure that checking if the tracker is stopped can be done
even if there is no tracker instance alive.