Fix symbol overriding for non-PIE Python static binaries
If the Python that is running `pensieve` is compiled statically and is
generating rellocations of type `R_X86_64_GLOB_DAT` for some symbols,
and those symbols are nor marked as WEAK by the original sources (this
happens in `glibc` with `malloc()` and `free()`), is possible that when
pensieve gets the address of the original symbols by executing
`&::symbol` then we will not get the address of the symbol at its source
but some PLT entry for the Python binary. This is because the linker
always sees the executable as its main scope and if the Python binary is
statically compiled with `--export-dynamic` (as normally is), then the
linker is free to choose to bind the symbol resolution against the
binary, even if those symbols are initially undefined (this happens with
`malloc()` and `free()`.

For this reason, getting the symbols via standard linker relocations is
not reliable so we need to fetch them manually by scanning all loaded
shared objects (and the executable). We scan the dynamic table of all
these objects searching for the symbols that we need. In the regular
case (when we have `libpython.so` or any of the previously mentioned
preconditions are false), the address that we are going to obtain is the
same as the one that we got via relocations, but in some cases this will
not be true.

When fetching the symbols from the dynamic table, we use the `DT_HASH`
and the `DT_GNU_HASH` hash tables to efficiently scan for those symbols
(the linker doesn't allow to do this in a non-efficient way, as there is
no simple way of getting the size of the dynamic table to do a linear
scan).