Handle invalid input files in reporters
When parsing the allocation records we don't currently check if the
input file is valid. This leads to a crash when we attempt to allocate
internal data structures with an invalid size read from the corrupt file
header.

To tackle this, add a magic number to the header and verify this during
parsing. We raise an exception during parsing if the magic value is
invalid, which the reporters can handle and display an error message.

Authored-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Reviewed-by: Pablo Galindo <pgalindo3@bloomberg.net>
Signed-off-by: László Kiss Kollár <lkisskollar@bloomberg.net>