Use `popstate` event to handle report navigation
Instead of relying on the `onhashchange` event, which can also be
triggered when the URL hash updates, we can use the `popstate` event
which only triggers when the history stack changes.

Note that the forward history does not get populated when navigating
within frames, so the only case we have to be concerned about is back
navigation.

Authored-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Reviewed-by: Pablo Galindo <pgalindo3@bloomberg.net>
Signed-off-by: László Kiss Kollár <lkisskollar@bloomberg.net>