Write extra metadata in the header
To enrich the generated reports with more information, write the
full command line used to generate the data and the start and end times
of the tracking into the binary file. Expose these on the `Tracker`
object as a `dict` for now. We will later revisit the API, this is just
to keep it simple for now.

Authored-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Reviewed-by: Pablo Galindo <pgalindo3@bloomberg.net>
Signed-off-by: László Kiss Kollár <lkisskollar@bloomberg.net>