Expose a `Metadata` class
Instead of exposing all of the tracking metadata through a dict, provide
a `Metadata` class with properties for each field. With this change we
also hide the file version field which has no value on the public API.

Note that the `Metadata` class is a dataclass defined in a separate
Python file. This is necessary as declaring the dataclass in Cython
produces a class which does not take kwargs in the initializer and has
other limitations. This was also done to keep the `.pyx` file small.