Use millisecond accuracy in start and end times
The current granularity of the start and end time attributes on the
metadata is seconds, which is not useful when a run completes in less
then a second.

The C++ layer stores milliseconds since epoch to avoid exposing the
`std::chrono` types to Cython and simplify the conversion to Python
`datetime` objects.