Reduce likelihood of line changes in .pyx file
Some of our tests rely on the line information in `_pensieve.pyx` and
every time we introduce a line change in this file, the tests break. To
reduce the likelihood of such breakage, move the test functions in
`_pensieve.pyx` to the top of the file, so that changes lower in the
core functionality will not affect these line numbers.