Expose peak memory usage in `Metadata`
We already know the peak memory usage as we compute it when we take the
high watermark snapshot, but we are not exposing this information.
This patch adds the peak memory in the `Metadata` class so that we
can display it in HTML reports.

In order to avoid recomputing the high watermark index every time we
need the peak memory or the index, we cache the results on the first
instance.

With this change we parse the file header in `__init__`. This is useful
as a check to fail early in case the input file is corrupt. This way we
don't have to worry about unhandled exceptions due to a bad file in
every method of the class.