Don't reset the tracking if the current thread is already being tracked
To allow tracking extensions that are spawning native threads and
calling `PyGILState_Ensure` to get into the interpreter we currently
hook into calls to `PyGILState_Ensure` and install the tracking
function, resetting any given state in case the thread was already
registered previously.

Unfortunately this logic doesn't work correctly if an extension calls
`PyGILState_Ensure` directly instead of calling `Py_BEGIN_ALLOW_THREADS`
or similar. This is totally correct (although unusual) and can be done
even inf the GIL is held at the time. We cannot reset the tracking state
in this case if the given thread is already being tracked as this is
just a regular function call.

To solve this, we need to maintain some state in the tracker that allows
us to identify if a given thread is being tracked or not so we can
decide to reset the tracking state accordingly.

Authored-by: pgalindo3 <pgalindo3@bloomberg.net>
Reviewed-by: Pradyun Gedam <pgedam@bloomberg.net>
Signed-off-by: pgalindo3 <pgalindo3@bloomberg.net>