Simplify checking for registered threads
To simplify how we handle the installation of the tracking function
within threads, we can access directly the storage location of the
tracing function withing the thread state. Technically this is not a
exposed interface but is exposed on the headers (is also used by Cython
and other projects). This means that this will potentially need some
versioned macros to adapt this to future code changing this field, but
this is unlikely and in that case, a stable C-API version to get this
field seems more likely.

Authored-by: pgalindo3 <pgalindo3@bloomberg.net>
Reviewed-by: lkisskol <lkisskollar@bloomberg.net>
Signed-off-by: pgalindo3 <pgalindo3@bloomberg.net>