Make life time of patched symbols fixed
Currently the container holding on to the patched symbols is a function
static variable. As this static might get destroyed while there are
still pending threads which can attempt to access the field, this can
lead to crashes.

This change wraps the symbol patching functions and the symbol container
into a class which is then initialized by the `Tracker`. This ensures
that the object will only get destroyed in the `Tracker`'s destructor.