Avoid TLS variables with non trivial destructors in tracking functions
Running the DMP test suite has reveled some problems with the
destruction of Thread Local Storage variables (TLS variables) that makes
`pensieve` crash or corrupt memory.

The problem relates to the fact that initializing TLS variables in a
compilation unit when/after `glibc` is tearing down a thread makes
`glibc` to destroy that variable when a new thread is created. This
causes a crash as when the destructor is triggered the associated memory
backing the TLS has already been freed/reused.

Another way to look at this problem is that once `glibc` has destroyed
`thread_local` variables with non-trivial destructors it will free them.
Since free calls into `pensieve`, we must prevent that from recreating
those thread_local variables with non-trivial destructors.

The solution to this problem is to use a TLS pointer (trivial
destructor) to access the last element of the vector for the Python
stack instead of using the TLS for the Python stack. In this way we
don't need to involve any destructor in the TLS variables getting
accessed (and potentially re-created) in the tracking functions.