Use bundling in `FrameCollector`
This allows separating the frames IDs for the frames created by
`RecordReader` and `RecordWriter`.

This decouples the `RecordReader` from the number of frames written in
the header, while is required for removing the number of frames from
the header, which makes it possible to write the header when tracking
begins.