Cache native allocations for FileReader queries
We can simplify the code and avoid re-parsing the allocations multiple
times if we avoid clearing the native allocations every time a query is
made.

Unlike the other query types, `get_allocation_records` can be lazily
evaluated and avoid using a large buffer, so this does not attempt to
pre-populate the native allocations. However, it has been changed to use
its own `RecordReader` instance instead of relying on the instance
variable.

Authored-by: lkisskol <lkisskollar@bloomberg.net>
Reviewed-by: pgalindo3 <pgalindo3@bloomberg.net>
Signed-off-by: lkisskol <lkisskollar@bloomberg.net>