Handle partial `munmap` deallocations
Currently when we process the high watermark snapshots, we consider the
deallocation's starting address when looking for existing allocations.
This does not work correctly with `munmap` which can deallocate an
arbitrary chunk of memory inside a larger `mmap`'d region.

To handle this, we introduce an interval tree to keep track of the
`mmap`'d intervals. With this, when processing the snapshots, we can
identify which `mmap` allocation record has been partially deallocated,
and only account for the actually freed memory. The interval tree uses a
vector internally to keep the implementation simple. We can revisit this
if it turns out to be a performance issue.

This code also handles situations when a single `munmap` deallocates
multiple contiguous `mmap`'d regions in one go.