Adaptatively adjust the number of native frames retrieved
When retrieving the native frames, we are using one of `libunwind`'s
fast APIs to fetch the instruction pointers. Unfortunately, this API
requires us to pass down a buffer with a fixed number of frames, but is
possible that there are more frames that the ones we requested. In this
case, we need to fall back to the slower but exact API to do the
unwinding. To avoid having to fall back constantly, we can adjust the
number of frames we request using the fast API when we realize that the
initial number is too small.