Add ability to track snapshots per-thread
In multithreaded applications it can be confusing to interpret the
reports when allocations from different threads are merged into a single
view. This change introduces a new `merge_thread` argument in snapshot
queries which will produce separate of records for different threads
instead of rolling them up into a single record.