Respect `--split-threads` on flamegraph reports
When `--split-threads` is used for flamegraph generation, identical
stack frames with differing thread IDs will be in separate records. The
flame graph reporter needs to take the thread ID into account when
reconstructing the stacks and treat stacks with differing thread IDs as
non-identical. We can achieve this simply by combining the frame tuple
with the thread ID in the stack frames `dict`.  In non-split thread
mode, we set the thread ID to -1 for all threads, thus the `dict` will
naturally treat all stacks associated with the same thread.