Include a flag in the header to indicate the presence of native traces
To correctly select what kind of traceback we need to display in the
different reporters automatically, we need to know if native traces
where requested when the tracking was done. By adding a new field to the
header of the result file we can query this information from the reader
and use it to our advantage.