Ensure that no allocations are registered while we dump new maps
When the application under tracing calls `dlopen` or `dlclose` we need
to dump all the memory mappings in order to inform the reader the
correct source of all the pointers that exist. A problem can happen if
allocations are registered while we are dumping the mappings because
this causes the reader to search for the source of the allocations in a
set of memory maps that is not yet complete. For this reason we need
exclusive access to the writer while we dump the memory mappings so we
need to add an extra lock that ensures that we cannot register
allocations while we invalidate the module cache.