Ensure the Tracker is deactivated in child process
We currently have a `fork()` handler that deactivates tracking
allocations in children but we currently do not deactivate the rest of
the tracker functionality which results in the children writing frames
to the same file and this manifests as duplicated frames or corrupted
result files.

To fix this we need to ensure that the entirety of the tracker is
deactivated in the child before exiting the fork handler.