Move label generation into JavaScript
Generating the label to display the number of allocations in the tool
tip on the flame graph is currently happening in Python. Because of
this, if we regenerate the data on the frontent, the labels don't get
updated.  Moving this logic into JS makes more sense, especially since
all other representation-related logic lives there.