Limit the depth of deep calls in the flame-graph reporter
When dumping the data to json, is very possible that we reach the
recursion limit of the Python interpreter if the original data contains
a very deep call. To avoid this, limit the number of children that we
link when constructing the graph and substitute anything deeper by a
"continuation" node.