Fix root node memory recalculation
When we filter the flame graph for a specific thread, the root node gets
recalculated to display only allocations which happened on the selected
thread. This calculation is incorrect, as it recursively walks all
children instead of taking only the direct descendants into account. As
the nodes already accumulate the allocation stats, we don't have to
recursively walk the subtrees.