Deactivate Cython profiling and tracing by default
Having Cython emit profiling frames for our internal machinery makes
controlling the Python stack much more complex than it should be, as we
need to deal with some exit frames that we never saw enter (like
`__exit__`). For this reason and to make the extension a bit more
performant, deactivate profiling and tracing by default and only
activate it where we want it (like for the test functions).