Ensure that the entry frame is unset if we return above the stack
When Python threads are executed, is possible that we will return above
the stack where the tracking function was activated, receiving `POP`
events for frames that we never saw being pushed. In this case, we need
to ensure that the entry frame is unset so we don't incorrectly query
it's line number in the `getCurrentPythonLineNumber()`. This currently
manifests in a (very rare) segmentation fault as the frame can be freed
once we return above it.