Correctly handle `EINTR` in `SocketSource`
If we get `EINTR` and then the remote hangs up, we will end up with an
infinite loop as `errno` is not reset and we will always hit that
condition in the loop. The alternative implementation correctly handles
the `EOF` event, and this also means that we can remove the current
`eof()` check.