Create socket-based reader and writer
Introduce `Sink` and `Source` interfaces to abstract away low-level read
and write operations and make it possible to provide different
implementations for the underlying transport.

This adds a new, socket-based implementation, which can be used to
stream records into a socket using a `SocketWriter`, and read those
records on the other end of the socket using a `SocketReader`.