Use value types to configure `Tracker` output
The `Tracker` object always owns the output stream, and reusing the file
or the socket output in multiple invocations of the same tracker is
never a valid operation.

Instead of using the `Writer` classes which abstract the output, use
simple value types which hold only the configuration for the different
output types. This avoids any issues around the life time of the objects
and makes it clear that when attempting to reuse them, they do not refer
to same resource, but to the same configuration.