Avoid deadlocks due to log statement in write call
We have a log statement in the `SocketSink::writeAll()` function which
calls back into Python. We can hit this line when we are trying to write
a record, and this log statement will invoke the Python trace function
again, causing another write on the `RecordReader`. This will block
forever as the class protects writes with a mutex.

For now, just substitute the log statement by writing to `cerr`. We can
consider throwing an exception with the error, or removing the message
and handling the error on a higher level later. We have plans to clean
up the interface, as right now some operations on the `Sink` classes
return errors, while others throw exceptions, and the return values are
not always handled.