Remove `reader` property from `Tracker`
This property was kept for convenience, as it allowed accessing the
output file on a `Tracker` instance. However, with the addition of the
socket output, this makes the API more complicated, as the `reader`
property cannot support providing access to the socket. It also makes it
possible to misuse the `Tracker` by creating an instance just to get the
`reader` attribute on it, which makes it impossible to open the output
for writing during the creating of the `Tracker` object, as we don't
know at that point whether the file will be used for reading, or
writing.