Explain what failed if output file creation fails
In a previous commit we removed the check in the `Tracker` to see if the
output file already exists, to rely on the C++ implementation to throw
an exception if the open call fails. This however is a more general
exception for any I/O error, and doesn't say what actually went wrong.

Include the human-readable version of `errno` in the exception text, so
that the error message captures what actually went wrong.