Drop the anonymous namespace
This made it impossible to `reduced_snapshot_map_t` in multiple
translation units, since the compiler sees the definitions across the
different translation units as distinct; due to this anonymous
namespace. This would result in a mismatch and functions that used
anything defined in this namespace as part of their signature would
present the warning "... used but never defined".

This works around that by exposing the contents of this anonymous
namespace directly in the `pensieve::api` namespace, which avoids the
definitions being viewed as distinct across translation units.