Implement `BackgroundSocketReader` and rework `SocketReader` to use it
This introduces a background worker thread, that receives the records
over the socket, moving the I/O for the socket into a separate thread.
This also exposes the ability to get a snapshot of the current heap, as
described by the records that have been received and processed by the
worker thread so far.

This functionality is necessary to unblock the main thread when it is
rendering the table for the live reporter, while the background worker
thread keeps receiving the records over the socket and aggregating them.