Move the creation of the thread into an explicit method
This eliminates a race condition, where the thread starts executing,
receives a record and passed that record into the aggregator before the
aggregator is initialized. This is possible since the thread can be
initialized (i.e. started) before the aggregator due to the order of
declaration of the data members for them in the class.

This also avoids potential issues due to the `vtable` being initialized
_after_ the constructor is executed, which could result in the thread
being passed an invalid/incorrect function pointer.

http://bburl/cpp-thread-class was a useful reference for this change.