Add a terminal tree reporter
To allow to quickly inspect a `pensieve` result file without the need to
generate HTML, add a new terminal-based tree reporter that is able to
print the biggest n allocations in a tree form. The tree reporter is
also able to collapse 1-node chains in the tree to allow the user to
visualize better the relationship between allocations.

Authored-by: pgalindo3 <pgalindo3@bloomberg.net>
Reviewed-by: László Kiss Kollár <lkisskollar@bloomberg.net>
Signed-off-by: pgalindo3 <pgalindo3@bloomberg.net>