Uncouple the data source from the view in the live reporter
Refactor the live reporter so the source of the data (the reader), the
data (the snapshot) and the view (the rich renderable) are uncoupled so
we can test/use every component in isolation. This not only simplifies
the test suite but also allows us to extend the driver loop to have more
data sources and uncouples the data update tick rate from the display
tick rate.