Handle I/O errors in `FileSource` constructor
We don't currently check the state of the stream after opening it in
the `FileSource` constructor. Throw an exception in this case.

Note that we perform a separate check in Cython in the `FileReader`
class to check if the file exists. Given that we now have a check in the
underlying C++ implementation which would catch this, we could remove
this check. However, due to Cython#4442, we have to have a reference to
`IOError` in the .pyx file, otherwise Cython does not generate correct
code for the exception (an import from `builtins` doesn't work either).
Because of this, I opted to keep the extra check. It doesn't hurt and it
produces a clearer error message when the file doesn't exist.