Present a better error message when socket connection is interrupted
When the process gets a SIGINT while the socket is waiting on an
`accept` call, the system call is interrupted and the C++ code raises an
`IoError`, which is converted to an `OSError` that can be caught in
Python.

Moving the tracker initialisation to be separate from the context entry
enables more narrow scoped exception handling for the raised exception.