Stop the live view when the remote disconnects
Instead of looping forever, detect when no more data is available form
the remote, stop updating live stats like elapsed time, and display a
message to the user.

Note that even though we don't get any more data, we are not stopping
Rich's UI thread to let users interact with the view.