Handle signal masking caused by Jenkins setup
Our Jenkins instance runs with SigIgn: 0000000001001007 as on today.
The second-to-last bit is 1, which means that the process ignores
`SIGINT`.

This behavior is also inherited by all sub-processes spawned
by Jenkins. This is suboptimal in this case, as this test is explicitly
testing that the handling of `SIGINT` is correct, which requires that
the `SIGINT` actually interrupts the subprocess.

Explicitly setting the signal handler to the default handler is a
workaround for ensuring that the signal is handled normally instead of
being ignored.