Handle end of stream in `BackgroundSocketReader`
The `SocketSource::read()` function doesn't currently handle the end of
stream event and this results in the background thread in
`BackgroundSocketReader` forever attempting to read records when the
remote disconnects.

This change adds `eof` return values in the `streambuf` implementation,
and adds handling for these on the call sites. We also introduce an
`is_active()` check on the `BackgroundSocketReader` so that clients can
determine whether more records will be available. We are exposing this
in a separate function instead of making
`Py_GetSnapshotAllocationRecords` return NULL, as the underlying
aggregator can still have useful data the client might want to access,
even after the client has disconnected.