Correctly handle very quick programs in the TUI
The TUI first checks if the reader is still active and then plots the
results. It can happen that by the time the TUI reaches this point, the
background thread in the reader has already finished. In this case, we
will never print any results. To tackle this, we can add state the TUI
class to track whether it's still active, and only draw until this state
is flipped to `False`, when the reader becomes inactive.