Make command line handling consistent in TUI
This moves the default logic for unset and empty command line strings
into the TUI class, which makes it consistent with how the `pid`
argument is handled.

This also changes how an empty command line string is handled: it will
now be shown as "???" instead of an empty string, to make it clearer
that the command line is empty.